Refactor NeoFaker.Text
